refactor(standards): move the config notes into note::config
