+++ crates/miden-standards/src/account/auth/network_account/note_allowlist.rs
@@ -250,7 +250,7 @@ mod tests {
 
         // The map's ordering is determined by the StorageMapKey, so compare as sets.
         let mut expected: BTreeSet<NoteScriptRoot> = original_roots.into_iter().collect();
-        expected.insert(crate::note::NetworkAccountConfigNote::script_root());
+        expected.insert(crate::note::config::NetworkAccountConfigNote::script_root());
         expected.insert(crate::note::FeeSponsorshipNote::script_root());
         let actual: BTreeSet<NoteScriptRoot> =
             allowlist.allowed_script_roots().iter().copied().collect();
