+++ crates/miden-standards/src/note/config/blocklist_config.rs
@@ -22,110 +22,125 @@ use miden_protocol::{Felt, Word};
 
 use crate::StandardsLib;
 use crate::note::NetworkAccountTarget;
-use crate::note::costs::{NoteConsumptionCost, PAUSE_CONFIG_CONSUMPTION_CYCLES};
+use crate::note::costs::{BLOCKLIST_CONFIG_CONSUMPTION_CYCLES, NoteConsumptionCost};
 
 // NOTE SCRIPT
 // ================================================================================================
 
-/// Path to the PAUSE_CONFIG note script procedure in the standards library.
-const PAUSE_CONFIG_SCRIPT_PATH: &str = "::miden::standards::notes::pause_config::main";
+/// Path to the BLOCKLIST_CONFIG note script procedure in the standards library.
+const BLOCKLIST_CONFIG_SCRIPT_PATH: &str = "::miden::standards::notes::blocklist_config::main";
 
-// Initialize the PAUSE_CONFIG note script only once.
-static PAUSE_CONFIG_SCRIPT: LazyLock<NoteScript> = LazyLock::new(|| {
+// Initialize the BLOCKLIST_CONFIG note script only once.
+static BLOCKLIST_CONFIG_SCRIPT: LazyLock<NoteScript> = LazyLock::new(|| {
     let standards_lib = StandardsLib::default();
-    let path = Path::new(PAUSE_CONFIG_SCRIPT_PATH);
+    let path = Path::new(BLOCKLIST_CONFIG_SCRIPT_PATH);
     NoteScript::from_package_reference(standards_lib.as_ref(), path)
-        .expect("Standards library contains PAUSE_CONFIG note script procedure")
+        .expect("Standards library contains BLOCKLIST_CONFIG note script procedure")
 });
 
-// PAUSE CONFIG
+// BLOCKLIST CONFIG
 // ================================================================================================
 
 /// A management action of the
-/// [`PausableManager`](crate::account::access::pausable::PausableManager) component that a
-/// [`PauseConfigNote`] triggers on the account that consumes it.
+/// [`BlocklistManager`](crate::account::policies::BlocklistManager) component that a
+/// [`BlocklistConfigNote`] triggers on the account that consumes it.
 ///
-/// The action is encoded into the note's storage (see [`NoteStorage`] conversion below) and is
-/// fixed at note creation, bound into the note commitment. The consuming account's
-/// `PausableManager` procedures authorize the action through the account-wide
+/// The action, together with its argument, is encoded into the note's storage (see [`NoteStorage`]
+/// conversion below) and is fixed at note creation, bound into the note commitment. The consuming
+/// account's `BlocklistManager` procedures authorize the action through the account-wide
 /// [`Authority`](crate::account::access::Authority) component.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PauseConfig {
-    /// Pause the account, blocking pause-gated procedures until a matching unpause.
-    Pause,
-    /// Unpause the account.
-    Unpause,
+pub enum BlocklistConfig {
+    /// Add `account` to the blocklist. Blocking an already blocked account is a noop.
+    BlockAccount { account: AccountId },
+    /// Remove `account` from the blocklist. Unblocking an account that is not blocked is a noop.
+    UnblockAccount { account: AccountId },
 }
 
-impl PauseConfig {
+impl BlocklistConfig {
     // SELECTORS
     // --------------------------------------------------------------------------------------------
 
     // Config note selectors stored in the first storage item. Keep in sync with
-    // `pause_config.masm`.
-    const SELECTOR_PAUSE: u8 = 0;
-    const SELECTOR_UNPAUSE: u8 = 1;
+    // `blocklist_config.masm`.
+    const SELECTOR_BLOCK_ACCOUNT: u8 = 0;
+    const SELECTOR_UNBLOCK_ACCOUNT: u8 = 1;
 
-    /// Returns the note storage values encoding this action, laid out as `[selector]`.
-    fn to_storage_values(self) -> Vec<Felt> {
+    /// Returns the selector encoding this action in the first storage item.
+    const fn selector(self) -> u8 {
         match self {
-            PauseConfig::Pause => vec![Felt::from(Self::SELECTOR_PAUSE)],
-            PauseConfig::Unpause => vec![Felt::from(Self::SELECTOR_UNPAUSE)],
+            BlocklistConfig::BlockAccount { .. } => Self::SELECTOR_BLOCK_ACCOUNT,
+            BlocklistConfig::UnblockAccount { .. } => Self::SELECTOR_UNBLOCK_ACCOUNT,
         }
+    }
+
+    /// Returns the account the action operates on.
+    const fn target(self) -> AccountId {
+        match self {
+            BlocklistConfig::BlockAccount { account }
+            | BlocklistConfig::UnblockAccount { account } => account,
+        }
+    }
+
+    /// Returns the note storage values encoding this action, laid out as `[selector,
+    /// account_suffix, account_prefix]`.
+    fn to_storage_values(self) -> Vec<Felt> {
+        let account = self.target();
+        vec![Felt::from(self.selector()), account.suffix(), account.prefix().as_felt()]
     }
 }
 
-impl From<PauseConfig> for NoteStorage {
-    fn from(config: PauseConfig) -> Self {
+impl From<BlocklistConfig> for NoteStorage {
+    fn from(config: BlocklistConfig) -> Self {
         NoteStorage::new(config.to_storage_values())
             .expect("number of storage items should not exceed max storage items")
     }
 }
 
-// PAUSE CONFIG NOTE
+// BLOCKLIST CONFIG NOTE
 // ================================================================================================
 
-/// A PauseConfig note: triggers a
-/// [`PausableManager`](crate::account::access::pausable::PausableManager) admin action on the
-/// account that consumes it.
+/// A BlocklistConfig note: triggers a
+/// [`BlocklistManager`](crate::account::policies::BlocklistManager) admin action on the account
+/// that consumes it.
 ///
 /// A single note script dispatches on a selector in the note's storage to one of the component's
-/// admin procedures (`pause`, `unpause`). Authorization is enforced by those procedures through
-/// the account-wide [`Authority`](crate::account::access::Authority) component, so the note carries
-/// no assets.
+/// admin procedures (`block_account`, `unblock_account`). Authorization is enforced by those
+/// procedures through the account-wide [`Authority`](crate::account::access::Authority) component,
+/// so the note carries no assets.
 ///
-/// The note is always public (for network execution) and tagged for `account` — the account
-/// carrying the `PausableManager` component whose pause state is being managed.
+/// The note is always public and tagged for `target` — the account carrying the
+/// `BlocklistManager` component whose blocklist is being managed.
 ///
-/// The note is bound to the target `account` by a
+/// The note is bound to `target` by a
 /// [`NetworkAccountTarget`](crate::note::NetworkAccountTarget) attachment: the script asserts
 /// that the consuming account matches that target before dispatching, so the note cannot be
 /// consumed by a third-party account that merely accepts its sender.
 ///
 /// The note must be public: the script rejects a non-public note. See
-/// [the module docs](crate::note#note-type-of-the-config-notes) for the layers that enforce it.
+/// [the module docs](crate::note::config#note-type) for the layers that enforce it.
 ///
-/// Construct one with the [builder](PauseConfigNote::builder); convert it into a protocol [`Note`]
-/// infallibly via `Note::from`.
+/// Construct one with the [builder](BlocklistConfigNote::builder); convert it into a protocol
+/// [`Note`] infallibly via `Note::from`.
 #[derive(Debug, Clone)]
-pub struct PauseConfigNote {
+pub struct BlocklistConfigNote {
     sender: AccountId,
     target: AccountId,
-    config: PauseConfig,
+    config: BlocklistConfig,
     serial_number: Word,
     attachments: NoteAttachments,
 }
 
 #[bon::bon]
-impl PauseConfigNote {
-    /// Builds a new [`PauseConfigNote`] that applies `config` to `account`.
+impl BlocklistConfigNote {
+    /// Builds a new [`BlocklistConfigNote`] that applies `config` to `target`.
     ///
     /// # Errors
     ///
     /// Returns an error if:
-    /// - `account` is not a public account (the note is bound to it via a `NetworkAccountTarget`,
+    /// - `target` is not a public account (the note is bound to it via a `NetworkAccountTarget`,
     ///   which requires a public target).
-    /// - the attachments carry a `NetworkAccountTarget` for an account other than `account`.
+    /// - the attachments carry a `NetworkAccountTarget` for an account other than `target`.
     /// - the attachments exceed their protocol limit (see [`NoteAttachments::new`]); the target
     ///   attachment occupies one of the available slots when the caller does not supply it.
     #[builder]
@@ -133,14 +148,14 @@ impl PauseConfigNote {
         #[builder(field)] mut attachments: Vec<NoteAttachment>,
         sender: AccountId,
         target: AccountId,
-        config: PauseConfig,
+        config: BlocklistConfig,
         serial_number: Word,
     ) -> Result<Self, NoteError> {
         // The note script asserts that the consuming account matches this target before
         // dispatching.
         NetworkAccountTarget::ensure_presence(&mut attachments, target).map_err(|err| {
             NoteError::other_with_source(
-                "failed to bind the PauseConfig note to its target account",
+                "failed to bind the BlocklistConfig note to its target account",
                 err,
             )
         })?;
@@ -157,24 +172,28 @@ impl PauseConfigNote {
     }
 }
 
-impl PauseConfigNote {
+impl BlocklistConfigNote {
     // CONSTANTS
     // --------------------------------------------------------------------------------------------
 
-    /// Number of storage items of a PauseConfig note: a single selector.
-    pub const NUM_STORAGE_ITEMS: usize = 1;
+    /// Number of storage items of a BlocklistConfig note: a selector followed by the account ID
+    /// the action operates on.
+    ///
+    /// Both actions carry the same arguments, so the layout is fixed at `[selector,
+    /// account_suffix, account_prefix]`.
+    pub const NUM_STORAGE_ITEMS: usize = 3;
 
     // PUBLIC ACCESSORS
     // --------------------------------------------------------------------------------------------
 
-    /// Returns the script of the PauseConfig note.
+    /// Returns the script of the BlocklistConfig note.
     pub fn script() -> NoteScript {
-        PAUSE_CONFIG_SCRIPT.clone()
+        BLOCKLIST_CONFIG_SCRIPT.clone()
     }
 
-    /// Returns the PauseConfig note script root.
+    /// Returns the BlocklistConfig note script root.
     pub fn script_root() -> NoteScriptRoot {
-        PAUSE_CONFIG_SCRIPT.root()
+        BLOCKLIST_CONFIG_SCRIPT.root()
     }
 
     /// Returns the account ID of the note's sender (the authorizing party under an owner- or
@@ -189,7 +208,7 @@ impl PauseConfigNote {
     }
 
     /// Returns the admin action carried by the note.
-    pub fn config(&self) -> PauseConfig {
+    pub fn config(&self) -> BlocklistConfig {
         self.config
     }
 
@@ -207,7 +226,7 @@ impl PauseConfigNote {
 // BUILDER EXTENSIONS
 // ================================================================================================
 
-impl<S: pause_config_note_builder::State> PauseConfigNoteBuilder<S> {
+impl<S: blocklist_config_note_builder::State> BlocklistConfigNoteBuilder<S> {
     /// Adds a single attachment to the note.
     pub fn attachment(mut self, attachment: impl Into<NoteAttachment>) -> Self {
         self.attachments.push(attachment.into());
@@ -224,15 +243,15 @@ impl<S: pause_config_note_builder::State> PauseConfigNoteBuilder<S> {
     }
 }
 
-impl<S: pause_config_note_builder::State> PauseConfigNoteBuilder<S>
+impl<S: blocklist_config_note_builder::State> BlocklistConfigNoteBuilder<S>
 where
-    S::SerialNumber: pause_config_note_builder::IsUnset,
+    S::SerialNumber: blocklist_config_note_builder::IsUnset,
 {
     /// Draws a serial number from `rng` and sets it on the builder.
     pub fn generate_serial_number(
         self,
         rng: &mut impl FeltRng,
-    ) -> PauseConfigNoteBuilder<pause_config_note_builder::SetSerialNumber<S>> {
+    ) -> BlocklistConfigNoteBuilder<blocklist_config_note_builder::SetSerialNumber<S>> {
         self.serial_number(rng.draw_word())
     }
 }
@@ -240,15 +259,15 @@ where
 // CONVERSIONS
 // ================================================================================================
 
-impl From<PauseConfigNote> for Note {
-    fn from(note: PauseConfigNote) -> Self {
-        // PauseConfig notes carry no assets and are always public for network execution; the action
-        // lives in the note storage.
+impl From<BlocklistConfigNote> for Note {
+    fn from(note: BlocklistConfigNote) -> Self {
+        // BlocklistConfig notes carry no assets and are always public for network execution; the
+        // action and its argument live in the note storage.
         let metadata = PartialNoteMetadata::new(note.sender, NoteType::Public)
             .with_tag(NoteTag::with_account_target(note.target));
         let recipient = NoteRecipient::new(
             note.serial_number,
-            PauseConfigNote::script(),
+            BlocklistConfigNote::script(),
             NoteStorage::from(note.config),
         );
 
@@ -259,9 +278,9 @@ impl From<PauseConfigNote> for Note {
 // NOTE CONSUMPTION COST
 // ================================================================================================
 
-impl NoteConsumptionCost for PauseConfigNote {
+impl NoteConsumptionCost for BlocklistConfigNote {
     fn consumption_cycles() -> u32 {
-        PAUSE_CONFIG_CONSUMPTION_CYCLES
+        BLOCKLIST_CONFIG_CONSUMPTION_CYCLES
     }
 }
 
@@ -283,20 +302,21 @@ mod tests {
 
     /// The builder produces a public, asset-less note tagged for the managed account.
     #[test]
-    fn builder_builds_pause_config_note() {
+    fn builder_builds_blocklist_config_note() {
         let mut rng = RandomCoin::new(Word::empty());
         let managed = account_id(1);
-        let sender = account_id(2);
+        let owner = account_id(2);
+        let blocked = account_id(3);
 
-        let note = PauseConfigNote::builder()
-            .sender(sender)
+        let note = BlocklistConfigNote::builder()
+            .sender(owner)
             .target(managed)
-            .config(PauseConfig::Pause)
+            .config(BlocklistConfig::BlockAccount { account: blocked })
             .generate_serial_number(&mut rng)
             .build()
             .unwrap();
 
-        assert_eq!(note.sender(), sender);
+        assert_eq!(note.sender(), owner);
         assert_eq!(note.target(), managed);
 
         let note = Note::from(note);
@@ -305,13 +325,35 @@ mod tests {
         assert_eq!(note.assets().num_assets(), 0);
     }
 
-    /// `Pause` / `Unpause` storage is a single selector item.
+    /// `BlockAccount` storage is `[selector, account_suffix, account_prefix]`.
     #[test]
-    fn action_storage_layout() {
-        let pause = NoteStorage::from(PauseConfig::Pause);
-        assert_eq!(pause.items(), &[Felt::from(PauseConfig::SELECTOR_PAUSE)]);
+    fn block_account_storage_layout() {
+        let blocked = account_id(3);
+        let storage = NoteStorage::from(BlocklistConfig::BlockAccount { account: blocked });
 
-        let unpause = NoteStorage::from(PauseConfig::Unpause);
-        assert_eq!(unpause.items(), &[Felt::from(PauseConfig::SELECTOR_UNPAUSE)]);
+        assert_eq!(
+            storage.items(),
+            &[
+                Felt::from(BlocklistConfig::SELECTOR_BLOCK_ACCOUNT),
+                blocked.suffix(),
+                blocked.prefix().as_felt(),
+            ]
+        );
+    }
+
+    /// `UnblockAccount` storage is `[selector, account_suffix, account_prefix]`.
+    #[test]
+    fn unblock_account_storage_layout() {
+        let blocked = account_id(3);
+        let storage = NoteStorage::from(BlocklistConfig::UnblockAccount { account: blocked });
+
+        assert_eq!(
+            storage.items(),
+            &[
+                Felt::from(BlocklistConfig::SELECTOR_UNBLOCK_ACCOUNT),
+                blocked.suffix(),
+                blocked.prefix().as_felt(),
+            ]
+        );
     }
 }
