+++ crates/miden-standards/src/note/config/faucet_metadata_config.rs
@@ -165,7 +165,7 @@ impl From<FaucetMetadataConfig> for NoteStorage {
 /// consumed by a third-party account that merely accepts its sender.
 ///
 /// The note must be public: the script rejects a non-public note. See
-/// [the module docs](crate::note#note-type-of-the-config-notes) for the layers that enforce it.
+/// [the module docs](crate::note::config#note-type) for the layers that enforce it.
 ///
 /// Construct one with the [builder](FaucetMetadataConfigNote::builder); convert it into a protocol
 /// [`Note`] infallibly via `Note::from`.
