+++ crates/miden-standards/src/note/config/min_burn_amount_config.rs
@@ -69,7 +69,7 @@ static MIN_BURN_AMOUNT_CONFIG_SCRIPT: LazyLock<NoteScript> = LazyLock::new(|| {
 /// switched in.
 ///
 /// The note must be public: the script rejects a non-public note. See
-/// [the module docs](crate::note#note-type-of-the-config-notes) for the layers that enforce it.
+/// [the module docs](crate::note::config#note-type) for the layers that enforce it.
 ///
 /// Construct one with the [builder](MinBurnAmountConfigNote::builder); convert it into a protocol
 /// [`Note`] infallibly via `Note::from`.
