+++ crates/miden-standards/src/note/config/rbac_config.rs
@@ -140,7 +140,7 @@ impl From<RbacConfig> for NoteStorage {
 /// execution.
 ///
 /// The note must be public: the script rejects a non-public note. See
-/// [the module docs](crate::note#note-type-of-the-config-notes) for the layers that enforce it.
+/// [the module docs](crate::note::config#note-type) for the layers that enforce it.
 ///
 /// Construct one with the [builder](RbacConfigNote::builder); convert it into a protocol [`Note`]
 /// infallibly via `Note::from`.
